update FAQ to tsx and update copies
