+++ src/app/FAQ/page.tsx
@@ -8,7 +8,7 @@ import FAQComponent from "@/app/_components/FAQCompent/FAQComponent";
 const FAQ:React.FC = () => {
     return (
         <div className="universe-bg px-4 h-full flex-1 flex flex-col pb-12 rounded-3xl">
-            <FAQComponent items={FAQData} linkColor='text-primary007' />
+            <FAQComponent items={FAQData} />
             <div className="w-full h-[82px] rounded-md bg-primary010 flex justify-start items-center mt-8">
                 <div className="w-12 h-12 relative flex items-center mr-3">
                     <Image src={'/svg/faq.svg'} layout='fill' objectFit='cover' alt={'faq'} />
