--- src/settings/FAQData.ts
@@ -1,26 +1,0 @@
-export const FAQData = [
-  {
-    question: 'What is Philosopher\'s Stone?',
-    answer: 'Philosopher\'s Stone is a unique platform that allows you to craft, send, and receive Gifts with delightful and mysterious experience. It integrates the excitement of Blind Box gifting with the use of [Spore Protocol](https://spore.pro), offering a secure and creative way to share your joy. '
-  },
-  {
-    question: 'What is a Blind Box?',
-    answer: 'A Blind Box is a type of gift where the contents are a surprise until it is opened. In Philosopher\'s Stone, you have the choice to send Gifts in a traditional, known format or opt for the Blind Box format, allowing for a more mysterious and thrilling gifting experience.'
-  },
-  {
-    question: 'How do I create a Gift or a Blind Box?',
-    answer: 'To create a Gift or a Blind Box, simply navigate to the "Create" tab from the top-right menu. Here, you can upload an image as your Gift and mint it using the corresponding CKB amount.'
-  },
-  {
-    question: 'How do I send a Gift or Blind Box?',
-    answer: 'Sending is easy: after creating your Gift or Blind Box, click on “Send as Gift” from its detail page. You can either enter the recipient\'s wallet address or choose to send via URL.'
-  },
-  {
-    question: 'How do I claim a Gift?',
-    answer: 'If you receive a Gift via a URL, you\'ll need to enter a passkey to claim it, ensuring secure and authorized access. For Gifts received directly on the platform, you can simply open and view them within Philosopher\'s Stone.'
-  },
-  {
-    question: 'What can I do with my Gift?',
-    answer: 'Gifts received on Philosopher\'s Stone can be stored in your digital inventory, shared with friends, or melted back into CKB, providing flexibility and value within the Nervos ecosystem.'
-  }
-];
